Updated system message and version name
In this update, the system message for the chat assistant has been modified to specify that all replies will be concise and under 200 characters. Additionally, the version name displayed in the console output has been changed from 'Jarvisnator' to 'Jarvis'. Minor formatting adjustments have also been made in the code.